MGMT-16041: Splits the preview badges into one component per module
