+++ libs/ui-lib/lib/common/components/ui/PreviewBadge.tsx
@@ -1,17 +1,9 @@
 import React from 'react';
 import { Label, Popover } from '@patternfly/react-core';
 import { InfoCircleIcon } from '@patternfly/react-icons';
-import { TECH_SUPPORT_LEVEL_LINK } from '../../config/constants';
 import ExternalLink from './ExternalLink';
 import { WithTestID } from '../../types';
 import { useTranslation } from '../../hooks/use-translation-wrapper';
-
-export type DeveloperPreviewProps = {
-  position?: PreviewBadgePosition;
-  className?: string;
-} & WithTestID;
-
-export type TechnologyPreviewProps = DeveloperPreviewProps;
 
 export enum PreviewBadgePosition {
   default,
@@ -19,13 +11,15 @@ export enum PreviewBadgePosition {
   inlineRight,
 }
 
-type PreviewBadgeProps = DeveloperPreviewProps & {
+export type PreviewBadgeProps = {
+  position?: PreviewBadgePosition;
+  className?: string;
   text: string;
   externalLink?: string;
   popoverText: string;
-};
+} & WithTestID;
 
-const PreviewBadge: React.FC<PreviewBadgeProps> = ({
+export const PreviewBadge: React.FC<PreviewBadgeProps> = ({
   position = PreviewBadgePosition.inline,
   className = 'pf-u-ml-md',
   text,
@@ -69,32 +63,4 @@ const PreviewBadge: React.FC<PreviewBadgeProps> = ({
     </span>
   );
 };
-
-export const DeveloperPreview: React.FC<DeveloperPreviewProps> = (props) => {
-  const { t } = useTranslation();
-  return (
-    <PreviewBadge
-      text={t('ai:Developer Preview')}
-      popoverText={t(
-        'ai:Developer preview features are not intended to be used in production environments. The clusters deployed with the developer preview features are considered to be development clusters and are not supported through the Red Hat Customer Portal case management system.',
-      )}
-      {...props}
-    />
-  );
-};
-DeveloperPreview.displayName = 'DeveloperPreview';
-
-export const TechnologyPreview: React.FC<TechnologyPreviewProps> = (props) => {
-  const { t } = useTranslation();
-  return (
-    <PreviewBadge
-      text={t('ai:Technology Preview')}
-      popoverText={t(
-        'ai:Technology preview features provide early access to upcoming product innovations, enabling you to test functionality and provide feedback during the development process.',
-      )}
-      externalLink={TECH_SUPPORT_LEVEL_LINK}
-      {...props}
-    />
-  );
-};
-TechnologyPreview.displayName = 'TechnologyPreview';
+PreviewBadge.displayName = 'PreviewBadge';
